// src/interfaces/Reserva.ts
export interface Reserva {
  id: number;
  from_name: string;
  email: string;
  whatsapp: string;
  user_type: string;
  activity_nature: string;
  project_details: string;
  target_audience: string;
  number_of_people: number;
  horario: string;
  stations: string;
  status: string;
  created_at: string;
  updated_at: string;
}